the Siv3D-Kun 2020 model.
It is the Siv3D-Kun 2020 model.
We recorded 4 loop animations: running, stopping, digging, and flying. There are 15 types of facial expression morphs.

It is the Siv3D-Kun 2020 model.
We recorded 4 loop animations: running, stopping, digging, and flying. There are 15 types of facial expression morphs.

The related files are the following four.
・Siv3DKun2.png
・Siv3DKun2.blend
・Siv3DKun2.glb
・Tools.png